requirements.txt
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
python-dotenv
reportlab
flask
python-docx
openpyxl
beautifulsoup4
lxml
unstructured
